move data converter/codec server components out of holocen
