+++ src/lib/components/data-encoder-status.svelte
